+++ src/ServiceProvider/DoctrineServiceProvider.php
@@ -3,32 +3,18 @@
 namespace OctoLab\Cilex\ServiceProvider;
 
 use Cilex\Application;
-use Cilex\Provider as Cilex;
+use Cilex\ServiceProviderInterface;
+use Doctrine\Common\EventManager;
+use Doctrine\DBAL\Configuration;
+use Doctrine\DBAL\DriverManager;
 use Doctrine\DBAL\Tools\Console\Helper\ConnectionHelper;
 use OctoLab\Common\Doctrine\Util\ConfigResolver;
 
 /**
  * @author Kamil Samigullin <kamil@samigullin.info>
- *
- * @see \Cilex\Provider\DoctrineServiceProvider
  */
-class DoctrineServiceProvider extends Cilex\DoctrineServiceProvider
+class DoctrineServiceProvider implements ServiceProviderInterface
 {
-    /** @var bool|string */
-    private $helperConnection;
-
-    /**
-     * @param bool|string $helperConnection to setup
-     * {@link \Doctrine\DBAL\Tools\Console\Helper\ConnectionHelper}, in particular for
-     * {@link \Doctrine\DBAL\Migrations\Tools\Console\Command\AbstractCommand}
-     *
-     * @api
-     */
-    public function __construct($helperConnection = false)
-    {
-        $this->helperConnection = $helperConnection;
-    }
-
     /**
      * @param Application $app
      *
@@ -39,32 +25,30 @@ class DoctrineServiceProvider extends Cilex\DoctrineServiceProvider
      */
     public function register(Application $app)
     {
-        parent::register($app);
-        if (isset($app['config']['doctrine']['dbal']['connections'])) {
-            $connections = $app['config']['doctrine']['dbal']['connections'];
-            if (isset($app['config']['doctrine']['dbal']['default_connection'])) {
-                $default = $app['config']['doctrine']['dbal']['default_connection'];
-                if (isset($connections[$default])) {
-                    $app['dbs.default'] = $default;
-                }
+        $config = $app->offsetGet('config');
+        if (!isset($config['doctrine'])) {
+            return;
+        }
+        ConfigResolver::resolve($config['doctrine:dbal']);
+        $app['connections'] = $app::share(function () use ($app, $config) {
+            $connections = new \Pimple();
+            foreach ($config['doctrine:dbal:connections'] as $id => $params) {
+                $connections->offsetSet(
+                    $id,
+                    DriverManager::getConnection($params, new Configuration(), new EventManager())
+                );
             }
-            $app['dbs.options'] = $connections;
-        }
-        if (isset($app['config']['doctrine']['dbal']['types'])) {
-            ConfigResolver::resolve($app['config']['doctrine']['dbal']);
-        }
-        if ($this->helperConnection) {
-            $dbs = $app->offsetGet('dbs');
-            if (is_bool($this->helperConnection)) {
-                $connection = $app->offsetGet('db');
-            } else {
-                $connection = $dbs[$this->helperConnection];
-            }
-            $app
-                ->offsetGet('console')
-                ->getHelperSet()
-                ->set(new ConnectionHelper($connection), 'connection')
-            ;
-        }
+            return $connections;
+        });
+        $app['connection'] = $app::share(function () use ($app, $config) {
+            $ids = $app['connections']->keys();
+            $default = $config['doctrine:dbal:default_connection'] ?: current($ids);
+            return $app['connections'][$default];
+        });
+        $app
+            ->offsetGet('console')
+            ->getHelperSet()
+            ->set(new ConnectionHelper($app->offsetGet('connection')), 'connection')
+        ;
     }
 }