+++ tests/ServiceProvider/MonologServiceProviderTest.php
@@ -22,13 +22,12 @@ class MonologServiceProviderTest extends TestCase
     {
         $app->register($this->getConfigServiceProviderForMonolog());
         $app->register(new MonologServiceProvider());
-        self::assertInstanceOf(Logger::class, $app['monolog']);
         self::assertInstanceOf(LoggerInterface::class, $app['logger']);
-        self::assertEquals($app['monolog'], $app['logger']);
         self::assertInstanceOf(Logger::class, $app['loggers']['app']);
         self::assertInstanceOf(Logger::class, $app['loggers']['debug']);
         self::assertInstanceOf(Logger::class, $app['loggers']['db']);
         self::assertEquals($app['logger'], $app['loggers'][$app['config']['monolog:default_channel']]);
+        self::assertEquals($app['console.name'], $app['logger']->getName());
     }
 
     /**
@@ -41,6 +40,24 @@ class MonologServiceProviderTest extends TestCase
     {
         try {
             $app->register(new MonologServiceProvider());
+            self::fail(sprintf('%s exception expected.', \InvalidArgumentException::class));
+        } catch (\InvalidArgumentException $e) {
+            self::assertTrue(true);
+        }
+    }
+
+    /**
+     * @test
+     * @dataProvider applicationProvider
+     *
+     * @param Application $app
+     */
+    public function registerEmpty(Application $app)
+    {
+        $app->register($this->getConfigServiceProvider('config', 'yml'));
+        $app->register(new MonologServiceProvider());
+        try {
+            $app->offsetGet('logger');
             self::fail(sprintf('%s exception expected.', \InvalidArgumentException::class));
         } catch (\InvalidArgumentException $e) {
             self::assertTrue(true);